Split out equations
